Add Qt4
The KDE Platform/Sdk will probably provide Qt5.

However some apps are still stuck with Qt4, so they will need to bundle
it.

Given how complicated it was to get it building, this seems like a good
thing to share.

Most of the optional features are disabled though, especially WebKit. If
those are needed, please use a new JSON manifest.

The Aarch64 patch comes from the Fedora package.