Revert "Updated ibus to version 1.5.20"
This reverts commit d0a6cbbad319b232e12a54288f30dcffcd62003d.

Since the GTK module is part of the ibus project it was being linked
against a newer ibus version than old runtimes have.

Ideally we should split ibus up a bit so the gtk module is separate
but for now just keep building an old version so new ABI doesn't
accidentally get pulled in.

Closes #69