Add SDL_gfx module
This requires special care and feeding to build on aarch64, which I
borrowed from org.manaplus.ManaPlus when org.tuxpaint.Tuxpaint gained a
dependency on this library.

I suspect (without proof) that this library is one of the reasons that
the other existing user of SDL_gfx on Flathub,
com.roguetemple.HyperRogue, is not built for aarch64.