libcanberra: Clean up GTK module
libcanberra is a standalone library on its own that applications can use
to play sounds. libcanberra-gtk is a thin wrapper around it to simplify
libcanberra usage from GTK 3 programs.

libcanberra also builds a GTK 3 module that can play event files on
events like every time you click. This is believed to be X11-only, and
not believed to be widely used.

Clean up the GTK 3 module, leaving only the actual library.